a lot of things changed :)

+++ DotNetShopping/DotNetShopping/Controllers/ApiController.cs
@@ -8,22 +8,18 @@ using System.Web.Mvc;
 
 namespace DotNetShopping.Controllers
 {
-
     public class ApiController : Controller
     {
-        // GET: Api
-
         private ApplicationDbContext db = new ApplicationDbContext();
-
-
+        // GET: Api
         [HttpPost]
         public ActionResult AddToCart(Int64 VariantId, int Qty)
         {
-            IEnumerable<CartListModel> model;
             if (User.Identity.IsAuthenticated)
             {
                 var UserId = User.Identity.GetUserId();
-                var cart = db.Carts.Where(x => x.UserId == UserId && x.VariantId == VariantId).FirstOrDefault();
+                var cart = db.Carts.Where(x => x.UserId == UserId &&
+                x.VariantId == VariantId).FirstOrDefault();
                 if (cart != null)
                 {
                     cart.Quantity += Qty;
@@ -39,56 +35,79 @@ namespace DotNetShopping.Controllers
                     db.SaveChanges();
                 }
                 CartListModel co = new CartListModel();
-
-                model = co.GetCart(UserId);
+                var model = co.GetCart(UserId);
                 return Json(new { Success = true, Cart = model });
-
             }
             return Json(new { Success = true });
-
         }
-
-
+        [HttpPost]
+        public ActionResult RemoveCart(Int64 VariantId)
+        {
+            if (User.Identity.IsAuthenticated)
+            {
+                var UserId = User.Identity.GetUserId();
+                var cart = db.Carts.Where(x => x.UserId == UserId &&
+                x.VariantId == VariantId).FirstOrDefault();
+                if (cart != null)
+                {
+                    db.Carts.Remove(cart);
+                    db.SaveChanges();
+                }
+                CartListModel co = new CartListModel();
+                var model = co.GetCart(UserId);
+                return Json(new { Success = true, Cart = model });
+            }
+            return Json(new { Success = true });
+        }
         public ActionResult GetShoppingCart()
         {
             var UserId = User.Identity.GetUserId();
             CartListModel co = new CartListModel();
             var model = co.GetCart(UserId);
             return Json(new { Success = true, Cart = model }, JsonRequestBehavior.AllowGet);
-
         }
 
-        
-
-        [HttpPost]
-        public ActionResult RemoveCart(Int64 variantId)
-        {
-            var UserId = User.Identity.GetUserId();
-            var cart = db.Carts.Where(x => x.VariantId == variantId && x.UserId == UserId).FirstOrDefault();
-            if (cart != null)
-            {
-                db.Carts.Remove(cart);
-                db.SaveChanges();
-
-            }
-            CartListModel co = new CartListModel();
-            var model = co.GetCart(UserId);
-            return Json(new { Success = true, Cart = model });
-        }
-
-        public ActionResult GetStatesFor (Int16 CountryId)
+        public ActionResult GetStatesFor(Int16 CountryId)
         {
             var states = db.States.Where(x => x.CountryId == CountryId)
                 .OrderBy(x => x.Name).ToList();
-            return Json(new { Success = true, States = states },JsonRequestBehavior.AllowGet);
+            return Json(new { Success = true, States = states }, JsonRequestBehavior.AllowGet);
+        }
+        
+        public ActionResult GetCitiesFor(Int16 CountryId, Int16 StateId)
+        {
+            var cities = db.Cities.Where(x => x.CountryId == CountryId && x.StateId == StateId)
+                .OrderBy(x => x.Name).ToList();
+            return Json(new { Success = true, Cities = cities }, JsonRequestBehavior.AllowGet);
+        }
+        public ActionResult GetShippingMethods(Int16 CountryId)
+        {
+            var model = db.ShippingMethods
+                .Join(db.ShippingCosts, sm => sm.ShippingMethodId, sc => sc.ShippingMethodId,
+                (sm, sc) => new { ShippingMethod = sm, ShippingCost = sc })
+                .Where(x => x.ShippingCost.CountryId == CountryId)
+                .Select( x => new ShippingListModel {
+                    ShippingMethodId = x.ShippingMethod.ShippingMethodId,
+                    Name = x.ShippingMethod.Name,
+                    Domestic = x.ShippingMethod.Domestic,
+                    International = x.ShippingMethod.International,
+                    CostHalf = x.ShippingCost.CostHalf,
+                    CostOne = x.ShippingCost.CostOne,
+                    CostOneHalf = x.ShippingCost.CostOneHalf,
+                    CostTwo = x.ShippingCost.CostTwo,
+                    CostTwoHalf = x.ShippingCost.CostTwoHalf,
+                    CountryId = x.ShippingCost.CountryId
+                })
+                .OrderBy(x => x.Name).ToList();
+            return Json(new { Success = true, ShippingMethods = model }, JsonRequestBehavior.AllowGet);
         }
 
-        public ActionResult GetCitiesFor (Int16 CountryId)
+        public ActionResult GetPaymentMethods(Int16 CountryId)
         {
-            var cities = db.Cities.Where(x => x.CountryId == CountryId)
-                .OrderBy(x => x.Name).ToList();
-            ViewBag.BillingCityId = new SelectList(cities);
-            return Json(new { Success = true, Cities = cities }, JsonRequestBehavior.AllowGet);
+            var isDomestic = CountryId == 2;
+            var model = db.PaymentMethods.Where(x => x.Domestic == isDomestic || x.International != isDomestic)
+                .OrderBy(x=> x.Name).ToList();
+            return Json(new { Success = true, PaymentMethods = model }, JsonRequestBehavior.AllowGet);
         }
     }
 }
